Add Data Sources section
A Data Sources section has been added to the digital-twin-safe.md file. This new section links to digital-twin-data-sources.md and provides users with a detailed list of data sources that can be used when creating a digital twin. Providing this information may facilitate users in their digital twin creation process.